// src/types/project.ts
export interface Project {
  id?: string;
  slug: string;
  title: string;
  description: string;
  category: string;
  overview: string;
  role: string;
  tools: string[];
  challenges: string[];
  solutions: string[];
  outcomes: string[];
  images: string[];
  coverImage: string;
  profileImage: string;
}